pyproject.toml
[tool.poetry]
name = "rag-from-scratch"
version = "0.1.0"
description = "A from-scratch RAG backend built step-by-step"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = "^0.115.1"
uvicorn = "^0.34.0"
openai = "^1.74.0"
python-dotenv = "^1.1.0"
httpx = "^0.28.1"
numpy = "^2.2.4"
psycopg2-binary = "^2.9.10"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
